requirements.txt
Flask
flask_cors
beautifulsoup4
requests
python-dotenv